1、 添加购物车逻辑 2。 订单逻辑和商品逻辑抽离

+++ YYG/src/com/yyg/model/OrderGroup.java
@@ -21,7 +21,7 @@ public class OrderGroup {
     public int price;
 
     @DatabaseField
-    public int createTime;
+    public long createTime;
 
     @DatabaseField(foreign = true)
     public User user;
